requirements.txt
flask
mysql-connector-python
fpdf
opencv-python
easyocr
numpy
pillow
gunicorn
